Create arch_auto_install.sh
This commit introduces a new script, arch_auto_install.sh, which aims to automate the process of installing Arch Linux on a system with specific configurations.

The script utilizes the Archinstall Python library to perform a guided installation of Arch Linux. It includes partitioning, formatting, and mounting the hard drive, selecting mirror servers, choosing profiles and additional packages, setting the hostname, adding a new user, and setting up the bootloader. The script will also check if multiple drives are present and prompt the user to select a drive for the installation.

Moreover, the script is configured to set up a system with specific storage partition sizes for the boot, root, and home directories. The Xorg server and Pipewire audio server are selected as default profiles, and no additional packages are installed.

This script is especially beneficial for users who frequently install Arch Linux, as it provides a convenient, automated way to complete the process. It also ensures consistent setup across different machines by using the same configuration parameters.

Please note that, while the script aims to make the installation process more straightforward, it is essential to understand each step thoroughly. Always review and modify the script according to your specific requirements before using it for an installation.